Finalizando o projeto

+++ src/views/users.php
@@ -24,7 +24,7 @@
                     <td><?= $user->end_date ?></td>
                     <td>
                         <a href="save_user.php?update=<?= $user->id ?>" class="btn btn-warning rounded-circle mr-2"><i class="icofont-edit"></i></a>
-                        <a href="?delete=<?= $user->id ?>" class="btn btn-danger rounded-circle"><i class="icofont-trash"></i></a>
+                        <a href="?delete=<?= $user->id ?>" class="btn btn-danger rounded-circle" onclick="return confirm('Deseja remover este usuário?');"><i class="icofont-trash"></i></a>
                     </td>
                 </tr>
             <?php endforeach; ?>
